Connexion

+++ src/main/kotlin/ktorm/Colis.kt
@@ -4,10 +4,10 @@ import org.ktorm.schema.*
 
 object Colis : Table<Nothing>("Colis") {
     val id = int("ID").primaryKey()
-    val code = varchar("Code")
-    val description = varchar("Description")
-    val poids = decimal("Poids") // ✅ Correction : suppression de la précision
-    val volume = decimal("Volume") // ✅ Correction : suppression de la précision
-    val dateReception = date("DateReception") // ✅ Garder `date()`
-    val idEmplacement = int("ID_Emplacement") // ✅ Clé étrangère
+    val code = varchar("Code") // ✅ Suppression du paramètre `length`
+    val description = varchar("Description") // ✅ Suppression du paramètre `length`
+    val poids = double("Poids") // ✅ Type correct pour `double`
+    val volume = double("Volume") // ✅ Type correct pour `double`
+    val dateReception = date("DateReception")
+    val idEmplacement = int("ID_Emplacement")
 }
